Rename package from py-load-chemb to py-load-spl.
This includes:
- Renaming the source directory in src/
- Updating pyproject.toml
- Updating imports and references in the code.
