--- src/py_load_chemb/__init__.py
@@ -1,1 +1,0 @@
-# This file makes the src/py_load_chemb directory a Python package.
