+++ src/py_load_spl/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the src/py_load_spl directory a Python package.
